# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Arinal Dzikrul - PKP Test",
        "version": "0.1"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "summary": "User Login",
                "operationId": "62f895d178d0c685b01f95155253fdeb",
                "requestBody": {
                    "description": "User credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "user1@gmail.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "summary": "User Register",
                "operationId": "2adf700bedff92413c61bdcb6fe033fd",
                "requestBody": {
                    "description": "Create User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "example": "User App"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "user1@gmail.com"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful register"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/auth/verify": {
            "get": {
                "summary": "verify user",
                "operationId": "92bf7e1b90c03ef58ec2dba273ac97ae",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "0": {
                            "bearerAuth": []
                        }
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}